+++ 1.8.9/src/main/java/io/github/axolotlclient/api/StatusUpdateProviderImpl.java
@@ -28,6 +28,7 @@ import java.util.Optional;
 import com.google.gson.JsonObject;
 import io.github.axolotlclient.api.requests.StatusUpdate;
 import io.github.axolotlclient.api.util.StatusUpdateProvider;
+import io.github.axolotlclient.modules.hypixel.HypixelGameType;
 import io.github.axolotlclient.modules.hypixel.HypixelLocation;
 import io.github.axolotlclient.util.GsonHelper;
 import io.github.axolotlclient.util.events.Events;
@@ -53,18 +54,15 @@ public class StatusUpdateProviderImpl implements StatusUpdateProvider {
 		if (entry != null) {
 
 			if (!entry.isLocal()) {
-				Optional<StatusUpdate.SupportedServer> optional = Arrays.stream(StatusUpdate.SupportedServer.values()).filter(s -> s.getAdress().matcher(entry.address).matches()).findFirst();
+				Optional<StatusUpdate.SupportedServer> optional = Arrays.stream(StatusUpdate.SupportedServer.values()).filter(s -> s.getAddress().matcher(entry.address).matches()).findFirst();
 				if (optional.isPresent()) {
 					StatusUpdate.SupportedServer server = optional.get();
 					if (server.equals(StatusUpdate.SupportedServer.HYPIXEL)) {
 						JsonObject object = HypixelLocation.get().thenApply(GsonHelper::fromJson).join();
-						StatusUpdate.GameType gameType = StatusUpdate.GameType.valueOf(object.get("gametype").getAsString());
+						HypixelGameType gameType = HypixelGameType.valueOf(object.get("gametype").getAsString());
 						String gameMode = getOrEmpty(object, "mode");
 						String map = getOrEmpty(object, "map");
-						int maxPlayers = Minecraft.getInstance().world.players.size();
-						int players = Minecraft.getInstance().world.players.stream()
-							.filter(e -> getGameMode(e) != WorldSettings.GameMode.CREATIVE && getGameMode(e) != WorldSettings.GameMode.SPECTATOR).mapToInt(value -> 1).reduce(0, Integer::sum);
-						return StatusUpdate.inGame(server, gameType.toString(), gameMode, map, players, maxPlayers);
+						return StatusUpdate.inGame(server, gameType.getName(), gameMode, map);
 					}
 				}
 			}