+++ 1.8.9/src/main/java/io/github/axolotlclient/modules/hud/gui/AbstractHudEntry.java
@@ -29,7 +29,6 @@ import com.mojang.blaze3d.platform.GlStateManager;
 import io.github.axolotlclient.AxolotlClientConfig.api.options.Option;
 import io.github.axolotlclient.AxolotlClientConfig.api.options.OptionCategory;
 import io.github.axolotlclient.AxolotlClientConfig.api.util.Colors;
-import io.github.axolotlclient.AxolotlClientConfig.impl.options.BooleanOption;
 import io.github.axolotlclient.AxolotlClientConfig.impl.options.DoubleOption;
 import io.github.axolotlclient.modules.hud.gui.component.HudEntry;
 import io.github.axolotlclient.modules.hud.util.DefaultOptions;
@@ -38,6 +37,7 @@ import io.github.axolotlclient.modules.hud.util.DrawUtil;
 import io.github.axolotlclient.modules.hud.util.Rectangle;
 import io.github.axolotlclient.util.ClientColors;
 import io.github.axolotlclient.util.Util;
+import io.github.axolotlclient.util.options.ForceableBooleanOption;
 import lombok.Getter;
 import lombok.Setter;
 import net.minecraft.client.Minecraft;
@@ -52,7 +52,8 @@ import net.minecraft.util.math.MathHelper;
 
 public abstract class AbstractHudEntry extends DrawUtil implements HudEntry {
 
-	protected final BooleanOption enabled = DefaultOptions.getEnabled();
+	@Getter
+	protected final ForceableBooleanOption enabled = DefaultOptions.getEnabled();
 	protected final DoubleOption scale = DefaultOptions.getScale(this);
 	protected final Minecraft client = Minecraft.getInstance();
 	private final DoubleOption x = DefaultOptions.getX(getDefaultX(), this);
@@ -148,7 +149,7 @@ public abstract class AbstractHudEntry extends DrawUtil implements HudEntry {
 		}
 		int scaledX = floatToInt(x.get().floatValue(), (int) Util.getWindow().getScaledWidth(), 0) - offsetTrueWidth();
 		int scaledY = floatToInt(y.get().floatValue(), (int) Util.getWindow().getScaledHeight(), 0)
-					  - offsetTrueHeight();
+			- offsetTrueHeight();
 		if (scaledX < 0) {
 			scaledX = 0;
 		}