+++ 1.8.9/src/main/java/io/github/axolotlclient/modules/tnttime/TntTime.java
@@ -26,9 +26,9 @@ import java.text.DecimalFormat;
 
 import io.github.axolotlclient.AxolotlClient;
 import io.github.axolotlclient.AxolotlClientConfig.api.options.OptionCategory;
-import io.github.axolotlclient.AxolotlClientConfig.impl.options.BooleanOption;
 import io.github.axolotlclient.AxolotlClientConfig.impl.options.IntegerOption;
 import io.github.axolotlclient.modules.AbstractModule;
+import io.github.axolotlclient.util.options.ForceableBooleanOption;
 import lombok.Getter;
 import net.minecraft.text.Formatting;
 import net.minecraft.text.LiteralText;
@@ -39,7 +39,7 @@ public class TntTime extends AbstractModule {
 
 	@Getter
 	private static final TntTime Instance = new TntTime();
-	public final BooleanOption enabled = new BooleanOption("enabled", false);
+	public final ForceableBooleanOption enabled = new ForceableBooleanOption("enabled", false);
 	private final OptionCategory category = OptionCategory.create("tnttime");
 	private final IntegerOption decimalPlaces = new IntegerOption("decimalplaces", 2, 0, 6);
 	private DecimalFormat format = new DecimalFormat("##");