+++ common/src/main/java/io/github/axolotlclient/api/Constants.java
@@ -29,7 +29,6 @@ public class Constants {
 	public final boolean ENABLED = true; // Completely disable these features and hide their existence
 	private final boolean DEVELOPMENT = false;
 	public final String API_URL = DEVELOPMENT ? "http://localhost:8000" : "https://api.axolotlclient.com";
-	public final String SOCKET_URL = DEVELOPMENT ? "ws://localhost:8000/gateway" : "wss://api.axolotlclient.com/gateway";
 	public final String PRIVACY_POLICY = "https://axolotlclient.com/privacy";
 	public final int STATUS_UPDATE_DELAY = 15; // The Delay between Status updates, in seconds. Discord uses 15 seconds so we will as well.
 	public final boolean TESTING = false; // When set to true, no requests will be sent